Finished the backend and frontend logic
Go to http://localhost:3000/auth to see the effect:
Unregistered Users: Users can create and manage plans which are associated with a temporary UUID stored in the browser's local storage. These plans persist even if the browser is refreshed.

Registered Users: When users register or log in, their existing plans associated with the temporary UUID are merged into their account. This allows them to access their plans from any device by logging in with their username.